requirements.txt
boto3==1.25.5
psycopg2-binary
pyOpenSSL